fix: LadderGameUtil >> GameUtil 이름변경 및 도메인 로직 제거
